requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
rest_framework_simplejwt
pytz